VWR and BWMA acreage
add posts for calculation of vegetation water requirements and a quick comparison of acreage for blackrock waterfowl management area between late Feb 2022 vs 2023. conclusion of the acreage was that BWMA exceeds 500 acres and not need to send staff out to GPS the flooded extent, which would be more important if the acreage was going to be under 500 acres by an unkown amount.